==> 04_rfmix/public_references_only/03.1_assess-rfmixQ-proportions.Rmd <==
---
title: '03.1_assess-rfmixQ-proportions.Rmd'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Setup 

```{r}
"%&%" <- function(a, b) paste0(a, b)
library("data.table")
library("tidyverse")
serv.dir <- "./"
work.dir.serv <- serv.dir %&% 
  "popgen/04_rfmix/public_references_only/"
input.dir <- work.dir.serv %&% "input_files/"
output.dir <- work.dir.serv %&% "output_files/"
tera.dir <- serv.dir %&% 
  "popgen/02.1_terastructure/"
pca.dir <- serv.dir %&% 
  "popgen/01_pca/public/bigsnpr_output/" 
```

```{r}
rgnkey.dir <- serv.dir %&% 
  "freeze_145k/03_rgc-qc-steps/input_files/"
rgnkey.file = rgnkey.dir %&% "freeze_145k_id-key.csv"
rgn.link.df <- fread(rgnkey.file)[, c(1, 3)]
names(rgn.link.df) <- c("MCPS.ID", "IID")
rgn.link.df$IID <-  as.character(rgn.link.df$IID)
```

```{r}
tera.df <- fread(tera.dir %&% "freeze_145k/output_files/" %&% 
                   "teraK12.inferred-ancestry-scores.txt") %>% 
  dplyr::select(one_of("IID", "afr.score", "amr.score", "eur.score"))

names(tera.df)[2:4] <- c("AFRICA", "AMERICA", "EUROPE")

pop.df <-  fread(pca.dir %&% "pc_projections_unrelateds_maf01_pop-info.txt")
names(pop.df)[1] <- "IID" #%>% 
pop.df <- dplyr::select(pop.df, one_of("IID", "region", "population"))
```


```{r}
rfmix.q.df <- fread(work.dir.serv %&% "output_files/"%&%
                      "global-ancestry-estimates.txt")
join.df <- inner_join(rfmix.q.df, rgn.link.df, by="IID")
mcps.sub.df <- filter(rfmix.q.df, IID %in% join.df$IID)
mcps.sub.df <- inner_join(mcps.sub.df, rgn.link.df, by="IID")
mcps.sub.df <- dplyr::select(mcps.sub.df, one_of("MCPS.ID", "AFRICA", 
                                    "AMERICA", "EUROPE"))
names(mcps.sub.df)[1] <- "IID"
ref.sub.df <- filter(rfmix.q.df, !(IID %in% join.df$IID))
rfmix.q.df <- rbind(mcps.sub.df, ref.sub.df)
```


```{r}
rfmix.q.df$IID <- purrr::map(rfmix.q.df$IID, function(s){
  ifelse(grepl("MCPS", s)==F, strsplit(s, split="-")[[1]][1], s)
}) %>% as.character(.)
```

```{r}
rfmix.q.df$IID[!(rfmix.q.df$IID %in% tera.df$IID)]
tera.df$IID[!(tera.df$IID %in% rfmix.q.df$IID)]
```

```{r}
rfmix.q.df <- filter(rfmix.q.df, IID %in% tera.df$IID)
tera.df <- filter(tera.df, IID %in% rfmix.q.df$IID)
```


# Functions 

```{r}
library("viridis")
reformat_df <- function(df){
  out.df <- c()
  pb <- txtProgressBar(min=0, max=dim(df)[1], style=3)
  for (i in 1:dim(df)[1]){
    setTxtProgressBar(pb, i)
    row.df <- df[i, ]
    prop.vec <- row.df[, (dim(row.df)[2]-3+1):dim(row.df)[2]] %>% as.numeric(.)
    grp.names <- row.df[, (dim(row.df)[2]-3+1):dim(row.df)[2]] %>% names(.)
    build.df <- data.frame("IID"=row.df$IID, 
                           "Proportion"=prop.vec, "Ancestry"=grp.names, 
                           stringsAsFactors = F)
    out.df <- rbind(out.df, build.df)
  }
  return(out.df)
}

pop_plot <- function(sub.df){
  ggplot(data=sub.df, aes(x=IID, y=Proportion)) + 
    geom_bar(stat="identity", 
             aes(fill=Ancestry, col=Ancestry)) + 
  scale_fill_manual(values=viridis(10)[c(2, 5, 8)]) + 
  scale_color_manual(values=viridis(10)[c(2, 5, 8)]) + 
    facet_wrap(~population, scales="free_x", 
               strip.position="bottom", 
               nrow=1) + 
    theme(axis.text.x=element_blank(), 
          axis.title.x = element_blank(), 
          axis.ticks.x=element_blank(), 
          panel.background = element_blank(), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          strip.placement = "outside", 
          strip.background = element_rect(fill="white"), 
          strip.text = element_text(size=6))
  
}

region_plot <- function(sub.df){
  ggplot(data=sub.df, aes(x=IID, y=Proportion)) + 
    geom_bar(stat="identity", 
             aes(fill=Ancestry, col=Ancestry)) + 
  scale_fill_manual(values=viridis(10)[c(2, 5, 8)]) + 
  scale_color_manual(values=viridis(10)[c(2, 5, 8)]) + 
    facet_wrap(~region, scales="free_x", 
               strip.position="bottom", 
               nrow=1) + 
    theme(axis.text.x=element_blank(), 
          axis.title.x = element_blank(), 
          axis.ticks.x=element_blank(), 
          panel.background = element_blank(), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          strip.placement = "outside", 
          strip.background = element_rect(fill="white"), 
          strip.text = element_text(size=6))
  
}
```

```{r}
tera.plt.df <- reformat_df(tera.df)
rfmixQ.plt.df <- reformat_df(rfmix.q.df)
```

```{r}
plt.df1 <- left_join(tera.plt.df, pop.df, by="IID") %>% na.omit(.)
plt.df2 <- left_join(rfmixQ.plt.df, pop.df, by="IID") %>% na.omit(.)
```

```{r}
lev.vec1 <- (filter(tera.plt.df, Ancestry=="EUROPE") %>% 
  arrange(desc(Proportion)))$IID
lev.vec2 <- (filter(rfmixQ.plt.df, Ancestry=="EUROPE") %>% 
  arrange(desc(Proportion)))$IID
plt.df1$IID <- factor(plt.df1$IID, levels=lev.vec1)
plt.df2$IID <- factor(plt.df2$IID, levels=lev.vec2)
```


## TeraStructure Plots 
```{r}
plt1a <- region_plot(filter(plt.df1, population!="MCPS"))
plt1b <- pop_plot(filter(plt.df1, region=="EUROPE"))
plt1c <- pop_plot(filter(plt.df1, region=="AMERICA", 
                         population!="MCPS"))
plt1d <- pop_plot(filter(plt.df1, population=="MCPS"))
ggsave(plot=plt1a, filename=work.dir.serv%&%
         "output_files/teraK.inferred-anc-prop.ref.png", 
       height=3, width = 6)
ggsave(plot=plt1b, filename=work.dir.serv%&%
         "output_files/teraK.inferred-anc-prop.EUR.png", 
       height=3, width = 12)
ggsave(plot=plt1c, filename=work.dir.serv%&%
         "output_files/teraK.inferred-anc-prop.AMR.png", 
       height=3, width = 12)
ggsave(plot=plt1d, filename=work.dir.serv%&%
         "output_files/teraK.inferred-anc-prop.MCPS.png", 
       height=3, width = 12)
```

## RFMix Plots 
```{r}
plt2a <- region_plot(filter(plt.df2, population!="MCPS"))
plt2b <- pop_plot(filter(plt.df2, region=="EUROPE"))
plt2c <- pop_plot(filter(plt.df2, region=="AMERICA", 
                         population!="MCPS"))
plt2d <- pop_plot(filter(plt.df2, population=="MCPS"))
ggsave(plot=plt2a, filename=work.dir.serv%&%
         "output_files/rfmix.inferred-anc-prop.ref.png", 
       height=3, width = 6)
ggsave(plot=plt2b, filename=work.dir.serv%&%
         "output_files/rfmix.inferred-anc-prop.EUR.png", 
       height=3, width = 12)
ggsave(plot=plt2c, filename=work.dir.serv%&%
         "output_files/rfmix.inferred-anc-prop.AMR.png", 
       height=3, width = 12)
ggsave(plot=plt2d, filename=work.dir.serv%&%
         "output_files/rfmix.inferred-anc-prop.MCPS.png", 
       height=3, width = 12)
```

## Correlation plots 

```{r}
library("reshape2")
library("viridis")

eval.df <- inner_join(tera.df, rfmix.q.df, by="IID")
mydata <- dplyr::select(eval.df, -one_of("IID"))
cormat <- round(cor(mydata), 2)
melted_cormat <- melt(cormat)
head(melted_cormat)
names(melted_cormat)[3] <- "Correlation"
tile.plt <- ggplot(data = melted_cormat,  
  aes(x=Var1,  y=Var2,  
      fill=Correlation, label=Correlation)) + 
  geom_tile() + 
  geom_text() + 
  scale_fill_viridis()
ggsave(plot=tile.plt, 
       filename=work.dir.serv%&%"output_files/teraS-rfmix-global-corr-tile.png", 
       height=5, width=8)

```

### Lookup MCPS samples to view more deeply in karyograms (04 scripts)

```{r}
rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AFRICA)) %>% head(2)
afr.vec <- (rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AFRICA)))$IID[1:2]
```

```{r}
rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(EUROPE)) %>% head(2)
eur.vec <- (rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(EUROPE)))$IID[1:2]
```

```{r}
rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AMERICA)) %>% head(2)
amr.vec <- (rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AMERICA)))$IID[1:2]
```


```{r}
set.seed(1)
amr.sub <- rfmix.q.df %>% filter(., grepl("MCPS", IID), AMERICA>=0.2, AMERICA<=0.8)
nrow <- dim(amr.sub)[1]
samp.df <-  amr.sub[sample(1:nrow, 10), ]
admix.vec <- samp.df$IID
example.vec <- c(afr.vec, eur.vec, amr.vec, admix.vec)
```


Save list 
```{r}
write.table(x=example.vec, file=work.dir.serv%&%
              "output_files/examples.n16.mcps-ids.txt", sep="\t", 
            quote=F, row.names=F, col.names=F)
rgn.vec <- c()
for (id in example.vec){
  iid <- filter(rgn.link.df, MCPS.ID==id)$IID
  rgn.vec <- c(rgn.vec, iid)
}
write.table(x=rgn.vec, file=work.dir.serv%&%
              "output_files/examples.n16.rgn-ids.txt", sep="\t", 
            quote=F, row.names=F, col.names=F)
```

Lookup up more AMR non-admixed samples 
```{r}
rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AMERICA)) %>% head(2)
amr.vec20 <- (rfmix.q.df %>% filter(., grepl("MCPS", IID)) %>% 
  arrange(., desc(AMERICA)))$IID[1:20]
write.table(x=amr.vec20, file=work.dir.serv%&%
              "output_files/examples.amr20.mcps-ids.txt", sep="\t", 
            quote=F, row.names=F, col.names=F)
rgn.vec <- c()
for (id in amr.vec20){
  iid <- filter(rgn.link.df, MCPS.ID==id)$IID
  rgn.vec <- c(rgn.vec, iid)
}
write.table(x=rgn.vec, file=work.dir.serv%&%
              "output_files/examples.amr20.rgn-ids.txt", sep="\t", 
            quote=F, row.names=F, col.names=F)
```
