==> 04_rfmix/public_references_only/03.1_rfmixQ-proportion-plot.Rmd <==
---
title: '03.1_assess-rfmixQ-proportions.Rmd'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Setup 

```{r}
"%&%" <- function(a,b) paste0(a,b)
library("data.table")
library("tidyverse")
serv.dir <- "./"
work.dir.serv <- serv.dir %&% 
  "popgen/04_rfmix/" %&%
  "public_references_only/"
input.dir <- work.dir.serv %&% "input_files/"
output.dir <- work.dir.serv %&% "output_files/"
pop.file <- serv.dir%&% "./shared/reference_datasets/" %&%
  "mais_information/reference-population-labels.txt"
```

```{r}
rgnkey.dir <- serv.dir %&% 
  "freeze_145k/03_rgc-qc-steps/input_files/"
rgnkey.file = rgnkey.dir %&% "freeze_145k_id-key.csv"
rgn.link.df <- fread(rgnkey.file)[,c(1,3)]
names(rgn.link.df) <- c("MCPS.ID","IID")
rgn.link.df$IID <-  as.character(rgn.link.df$IID)
```

```{r}
rfmix.q.df <- fread(work.dir.serv %&% "output_files/"%&%
                      "global-ancestry-estimates.txt")
join.df <- inner_join(rfmix.q.df,rgn.link.df,by="IID")
mcps.sub.df <- filter(rfmix.q.df,IID %in% join.df$IID)
mcps.sub.df <- inner_join(mcps.sub.df,rgn.link.df,by="IID")
mcps.sub.df <- dplyr::select(mcps.sub.df,one_of("MCPS.ID","AFRICA",
                                    "AMERICA","EUROPE"))
names(mcps.sub.df)[1] <- "IID"
ref.sub.df <- filter(rfmix.q.df,!(IID %in% join.df$IID))
rfmix.q.df <- rbind(mcps.sub.df,ref.sub.df)
```

```{r}
ref.sub.df <- filter(rfmix.q.df,!(grepl("MCPS",IID)))
mcps.sub.df <- filter(rfmix.q.df,(grepl("MCPS",IID)))
ref.id.vec <- purrr::map(ref.sub.df$IID,function(s){
  li <- strsplit(s,split="-")[[1]]
  paste0(li[2:length(li)],collapse="-")
}) %>% as.character(.)
ref.sub.df$IID <- ref.id.vec
mcps.sub.df <- mcps.sub.df
rfmix.q.df <- rbind(mcps.sub.df,ref.sub.df)
```

Save to save time in future
```{r}
write_rds(x=rfmix.q.df,file=work.dir.serv%&%"output_files/rfmix.q.df.RDS")
```


```{r}
pop.df <- fread(pop.file)
ref.df <- c()
pb <- txtProgressBar(min=0,max=dim(rfmix.q.df)[1],style=3)
for (i in 1:dim(rfmix.q.df)[1]){
  setTxtProgressBar(pb,i)
  samp <- rfmix.q.df$IID[i]
  sub.df <- filter(pop.df,sample==samp)
  if (dim(sub.df)[1]==0){
    sub.df <- data.table("sample"=samp,"population"="MCPS",
                         "region"="AMERICA",stringsAsFactors=F)
  }
  ref.df <- rbind(ref.df,sub.df)
}
names(ref.df)[1] <- "IID"
```

Save to save time in future
```{r}
write_rds(x=ref.df,file=work.dir.serv%&%"output_files/ref.df.RDS")
```

# Functions 

```{r}
library("viridis")
reformat_df <- function(df,k=3){
  out.df <- c()
  pb <- txtProgressBar(min=0,max=dim(df)[1],style=3)
  for (i in 1:dim(df)[1]){
    setTxtProgressBar(pb,i)
    row.df <- df[i,]
    prop.vec <- row.df[,(dim(row.df)[2]-k+1):dim(row.df)[2]] %>% as.numeric(.)
    grp.names <- row.df[,(dim(row.df)[2]-k+1):dim(row.df)[2]] %>% names(.)
    build.df <- data.frame("IID"=row.df$IID,
                           "Proportion"=prop.vec,"Ancestry"=grp.names,
                           stringsAsFactors = F)
    out.df <- rbind(out.df,build.df)
  }
  return(out.df)
}

pop_plot <- function(sub.df,col.vec,hide.text=TRUE,hide.legend=FALSE){
  plt <- ggplot(data=sub.df,aes(x=IID,y=Proportion)) + 
    geom_bar(stat="identity",
             aes(fill=Ancestry,col=Ancestry)) + 
  scale_y_continuous(breaks=seq(0,1,0.1)) + 
  scale_fill_manual(values=col.vec) + 
  scale_color_manual(values=col.vec) + 
    facet_wrap(~population,scales="free_x",
               strip.position="bottom",
               nrow=1) + 
    theme(axis.text.x=element_blank(),
          axis.title.x = element_blank(),
          axis.ticks.x=element_blank(),
          panel.background = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          strip.placement = "outside",
          strip.background = element_rect(fill="white"),
          strip.text = element_text(size=6))
  if (hide.text==TRUE){
    plt <- plt + theme(axis.text=element_blank(),
                       axis.title=element_blank(),
                       axis.ticks=element_blank())
  }
  if (hide.legend==TRUE){
    plt <- plt + theme(legend.position ="none")
  }
  return(plt)
}

region_plot <- function(sub.df,col.vec,hide.legend=FALSE){
  plt <- ggplot(data=sub.df,aes(x=IID,y=Proportion)) + 
    geom_bar(stat="identity",
             aes(fill=Ancestry,col=Ancestry)) + 
  scale_y_continuous(breaks=seq(0,1,0.1)) + 
  scale_fill_manual(values=col.vec) + 
  scale_color_manual(values=col.vec) + 
    facet_wrap(~region,scales="free_x",
               strip.position="bottom",
               nrow=1) + 
    theme(axis.text.x=element_blank(),
          axis.title.x = element_blank(),
          axis.ticks.x=element_blank(),
          panel.background = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          strip.placement = "outside",
          strip.background = element_rect(fill="white"),
          strip.text = element_text(size=6))
  if (hide.legend==TRUE){
    plt <- plt + theme(legend.position ="none")
  }
  return(plt)
}
```

```{r}
rfmixQ.plt.df <- reformat_df(rfmix.q.df)
```

Save to save time in future
```{r}
write_rds(x=rfmixQ.plt.df,file=work.dir.serv%&%
            "output_files/rfmixQ.plt.df.RDS")
```


```{r}
plt.df <- left_join(rfmixQ.plt.df,ref.df,by="IID") 
```


```{r}
lev.vec <- (filter(rfmixQ.plt.df,Ancestry=="EUROPE") %>% 
  arrange(desc(Proportion)))$IID %>% unique(.)
plt.df$IID <- factor(plt.df$IID,levels=lev.vec)
```

```{r}
plt.df$Ancestry <- factor(plt.df$Ancestry,
  levels=c("AFRICA","EUROPE","AMERICA"))
```

## RFMix Plots 
```{r}
library("cowplot")
vir.vec <- viridis(20)
cvec <- c(vir.vec[20],vir.vec[8],vir.vec[15])
plt1a <- region_plot(filter(plt.df,population!="MCPS"),
                     col.vec=cvec,hide.legend=TRUE)

plt1b <- pop_plot(filter(plt.df,population=="MCPS"),col.vec=cvec,hide.text=T)
mcps.plt <- pop_plot(filter(plt.df,population=="MCPS"),col.vec=cvec,
                     hide.legend = F,hide.text=F)

plt1 <- cowplot::plot_grid(plt1a,plt1b,nrow=1,rel_widths = c(1,5))
mcps.plt <- cowplot::plot_grid(mcps.plt,nrow=1)

```


Save plots
```{r}
svg(file=local.dir%&%"plots/rfmix.admixture-plot-freeze150k.svg",
    height =2,width = 12)
mcps.plt
dev.off()
```

## District-level MCPS plots 
Lookup MPCS district-specific samples 
```{r}
mcps.ref.df <- fread(serv.dir%&%"projects/mcps/data/phenotypes/" %&%
        "BASELINE.csv",header=T) %>% 
  dplyr::select(.,one_of("REGISTRO","COYOACAN","IZTAPALAPA"))
mcps.ref.df$district <- purrr::map(mcps.ref.df$COYOACAN,function(s){
  ifelse(s==1,"Coyoacan","Iztapalapa")
}) %>% as.character(.) # about a 4:6 ratio for plotting 
rgn.link.df <- fread(serv.dir%&%"projects/mcps/data/phenotypes/" %&%
                       "RGN_LINK.csv",header=T)
mcps.ref.df <- dplyr::inner_join(mcps.ref.df,rgn.link.df,
                                 by="REGISTRO")
mcps.sub.df <- filter(ref.df,population=="MCPS")
pb <- txtProgressBar(min=0,max=dim(mcps.sub.df)[1],style=3)
mcps.sub.df$district <- purrr::map(1:dim(mcps.sub.df)[1],function(i){
  setTxtProgressBar(pb,i)
  s <- (mcps.sub.df$IID[i] %>% strsplit(.,split="_"))[[1]][2]
  filter(mcps.ref.df,PATID==s)$district
}) %>% as.character(.)
rm(mcps.ref.df)
district.df <- mcps.sub.df
```

```{r}
district.df <- read_rds(file="popgen/04_rfmix/" %&%
  "output_files/district.df.RDS")
```


```{r}
library("cowplot")
vir.vec <- viridis(20)
##cvec <- c("#FDE725FF","#238A8DFF","#94D840FF")
cvec <- c(vir.vec[20],vir.vec[8],vir.vec[15])
plt2a <- region_plot(filter(plt.df,population!="MCPS"),
                     col.vec=cvec,hide.legend=TRUE)
plt2b <- pop_plot(filter(plt.df,population=="MCPS",
        IID %in% filter(district.df,district=="Coyoacan")$IID),
        col.vec=cvec,hide.text=T,hide.legend = TRUE)
plt2c <- pop_plot(filter(plt.df,population=="MCPS",
        IID %in% filter(district.df,district=="Iztapalapa")$IID),
        col.vec=cvec,hide.text=T,hide.legend = FALSE)
plt2 <- cowplot::plot_grid(plt2a,plt2b,plt2c,nrow=1,rel_widths = c(1,2,3))

coy.plt <- pop_plot(filter(plt.df,population=="MCPS",
        IID %in% filter(district.df,district=="Coyoacan")$IID),
        col.vec=cvec,hide.text=F,hide.legend = TRUE)
izt.plt <- pop_plot(filter(plt.df,population=="MCPS",
        IID %in% filter(district.df,district=="Iztapalapa")$IID),
        col.vec=cvec,hide.text=F,hide.legend = FALSE)
mcps.plt2 <- cowplot::plot_grid(coy.plt,izt.plt,nrow=1,rel_widths = c(4,6))

```

Save plots
```{r}
local.dir <- "popgen/" %&% "04_rfmix/public_references_only/"
ggsave(plot=mcps.plt2,filename=local.dir%&%
         "plots/rfmix.admixture-plot-freeze150k_districts.png",
       height=2,width=12)
```

```{r}
svg(file=local.dir%&%"plots/rfmix.admixture-plot-freeze150k_districts.svg",
    height=2,width=12)
mcps.plt2
dev.off()
```
